requirements.txt
Flask==2.3.2
twilio==8.12.1
openai==0.27.8
requests==2.31.0
apscheduler==3.10.1
python-dotenv==1.0.0
feedparser==6.0.11
gunicorn==21.2.0